+++ packages/backend/src/models/repositories/note.ts
@@ -13,7 +13,12 @@ import {
 	Notes,
 } from "../index.js";
 import type { Packed } from "@/misc/schema.js";
-import { countReactions, decodeReaction, nyaify } from "backend-rs";
+import {
+	countReactions,
+	decodeReaction,
+	nyaify,
+	shouldNyaify,
+} from "backend-rs";
 import { awaitAll } from "@/prelude/await-all.js";
 import type { NoteReaction } from "@/models/entities/note-reaction.js";
 import {
@@ -285,7 +290,13 @@ export const NoteRepository = db.getRepository(Note).extend({
 			lang: note.lang,
 		});
 
-		if (packed.user.isCat && packed.user.speakAsCat && packed.text) {
+		if (
+			packed.user.isCat &&
+			packed.user.speakAsCat &&
+			packed.text != null &&
+			meId != null &&
+			(await shouldNyaify(meId))
+		) {
 			const tokens = packed.text ? mfm.parse(packed.text) : [];
 			function nyaifyNode(node: mfm.MfmNode) {
 				if (node.type === "quote") return;